chore!: move components to proper repositories
Moves components to proper repositories, in accordance with the Gamut Audit notion
