+++ packages/gamut-labs/src/AppBar/index.tsx
@@ -1,7 +1,7 @@
+import { ContentContainer } from '@codecademy/gamut';
 import cx from 'classnames';
 import React from 'react';
 
-import { ContentContainer } from '../ContentContainer';
 // eslint-disable-next-line gamut/no-css-standalone
 import styles from './styles/index.module.scss';
