+++ packages/gamut-labs/src/CurriculumCard/index.tsx
@@ -4,13 +4,13 @@ import {
   CardProps,
   FlexBox,
   HeadingTags,
-  ProLabel,
   Text,
 } from '@codecademy/gamut';
 import { pxRem, theme } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
 import React, { useMemo } from 'react';
 
+import { ProLabel } from '..';
 import { TagColor } from './BottomTag/index';
 import { Footer } from './Footer/index';
 import { Image } from './Image/index';
