+++ packages/gamut-labs/src/GlobalPage/index.tsx
@@ -1,13 +1,12 @@
 import {
   AppWrapper,
-  Banner,
-  BannerProps,
   SkipToContent,
   SkipToContentTarget,
 } from '@codecademy/gamut';
 import { Background } from '@codecademy/gamut-styles';
 import React, { ComponentProps, forwardRef } from 'react';
 
+import { Banner, BannerProps } from '../Banner';
 import { GlobalFooter, GlobalFooterProps } from '../GlobalFooter';
 import { GlobalHeader, GlobalHeaderProps } from '../GlobalHeader';
