+++ packages/gamut-labs/src/ProLabel/index.tsx
