+++ packages/gamut/package.json
@@ -61,6 +61,7 @@
     "@emotion/jest": "^11.3.0",
     "@testing-library/dom": "^8.11.1",
     "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
     "@types/enzyme": "3.10.7",
     "enzyme": "3.11.0",
     "onchange": "^7.0.2"
