+++ packages/gamut/src/Drawer/index.tsx
@@ -1,8 +1,9 @@
-import { Box, BoxProps } from '@codecademy/gamut';
 import { breakpoints, timingValues } from '@codecademy/gamut-styles';
 import { AnimatePresence, motion } from 'framer-motion';
 import React from 'react';
 import { useMedia } from 'react-use';
+
+import { Box, BoxProps } from '../Box';
 
 const DrawerBase = motion(Box);
