+++ packages/styleguide/stories/Atoms/Drawer.stories.mdx
@@ -12,7 +12,7 @@ import { LinkTo } from '~styleguide/blocks';
   parameters={{
     subtitle:
       'Animated container that expands horizontally to show its contents.',
-    source: 'gamut-labs',
+    source: 'gamut',
     status: 'updating',
   }}
 />
