+++ packages/styleguide/stories/Brand Labs/Atoms/BetaSticker.stories.mdx
@@ -8,7 +8,7 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
   component={BetaSticker}
   parameters={{
     subtitle: `This Beta Sticker can be used for designating beta status in new features.`,
-    source: 'gamut',
+    source: 'gamut-labs',
     status: 'current',
   }}
   args={{}}
