+++ packages/styleguide/stories/Brand Labs/Atoms/Logos/Logo.stories.mdx
@@ -9,7 +9,7 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
   parameters={{
     subtitle: `Our proud logo. There is no a in the middle. Use sparingly, and make sure
 it has a lot of visual distinction to stand out appropriately.`,
-    source: 'gamut',
+    source: 'gamut-labs',
     design: {
       type: 'figma',
       url:
