+++ packages/styleguide/stories/Brand Labs/Atoms/Logos/ProLabel.stories.mdx
@@ -8,7 +8,7 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
   component={ProLabel}
   parameters={{
     subtitle: `This Pro logo will be used in the most recent designs for Curriculum Cards and Course overview pages.`,
-    source: 'gamut',
+    source: 'gamut-labs',
     status: 'current',
     design: {
       type: 'figma',
