+++ packages/styleguide/stories/Brand Labs/Molecules/Banner.stories.mdx
@@ -9,7 +9,7 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
   parameters={{
     subtitle: `Full-width banner to indicate important messaging at the top of a page or section.`,
     status: 'current',
-    source: 'gamut',
+    source: 'gamut-labs',
     design: {
       type: 'figma',
       url: 'https://www.figma.com/file/wTwIV8YfAgZ8eNpKSEd0Gn',
