+++ packages/styleguide/stories/Brand Labs/Molecules/Interstitial.stories.mdx
@@ -1,4 +1,4 @@
-import { ButtonDeprecated, Interstitial } from '@codecademy/gamut/src';
+import { ButtonDeprecated, Interstitial } from '@codecademy/gamut-labs/src';
 import title from '@codecademy/macros/lib/title.macro';
 import { PropsTable } from '@codecademy/storybook-addon-variance';
 import styled from '@emotion/styled';
@@ -13,7 +13,7 @@ Large content area used as a break between focused content, such as within
 a full-screen [Overlay](?path=/docs/atoms-overlay--overlay) in the LE between lessons.
 `,
     status: 'deprecated',
-    source: 'gamut',
+    source: 'gamut-labs',
   }}
   args={{
     children: `Place some happy text as the content of this interstitial for the
