+++ packages/styleguide/stories/Brand Labs/Molecules/Loaders/Shimmer.stories.mdx
@@ -8,7 +8,7 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
   component={Shimmer}
   parameters={{
     subtitle: 'Shimmer effect for boxes',
-    source: 'gamut',
+    source: 'gamut-labs',
     status: 'current',
     design: {
       type: 'figma',
