+++ packages/styleguide/stories/Brand Labs/Molecules/NotificationList.stories.mdx
@@ -98,7 +98,7 @@ export const notifications = [
   component={NotificationList}
   parameters={{
     subtitle: '',
-    source: 'gamut',
+    source: 'gamut-labs',
     status: 'current',
     design: {
       type: 'figma',
