+++ packages/styleguide/stories/Brand Labs/Molecules/Video.stories.mdx
@@ -10,7 +10,7 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
     subtitle: `
 This is an example video player that can play both youtube and vimeo videos
 given a video URL`,
-    source: 'gamut',
+    source: 'gamut-labs',
   }}
   args={{
     videoUrl: 'https://vimeo.com/145702525',
