+++ packages/styleguide/stories/Molecules/Breadcrumbs.stories.mdx
@@ -22,7 +22,7 @@ import { Canvas, Meta, Props, Story } from '@storybook/addon-docs/blocks';
   }}
   parameters={{
     subtitle: `Breadcrumb-style links for upwards navigation`,
-    source: 'gamut-labs',
+    source: 'gamut',
   }}
 />
