+++ packages/styleguide/stories/Molecules/Coachmark/index.stories.mdx
@@ -11,7 +11,7 @@ import { CoachmarkExample } from './Coachmark.examples';
   parameters={{
     subtitle:
       'A Coachmark helps bring attention to a CTA by showing a popover around that element.',
-    source: 'gamut-labs',
+    source: 'gamut',
     status: 'updating',
     design: {
       type: 'figma',
