+++ packages/styleguide/stories/Molecules/Flyout.stories.mdx
@@ -19,7 +19,7 @@ import { LinkTo } from '~styleguide/blocks';
   parameters={{
     subtitle:
       'Flyout is a general purpose container that animates into view. On button click, a container animates in from the left side of the window. It can be configured to close by pressing the "X" close icon, pressing they "esc" key, and/or clicking outside of the container.',
-    source: 'gamut-labs',
+    source: 'gamut',
     status: 'updating',
   }}
 />
