+++ packages/styleguide/stories/Molecules/Header.stories.mdx
@@ -9,7 +9,7 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
   parameters={{
     subtitle: `Header navigation bar and tabs with configurable areas.`,
     status: 'deprecated',
-    source: 'gamut-labs',
+    source: 'gamut',
     design: {
       type: 'figma',
       url: 'https://www.figma.com/file/TemYhRWOavsERWprDzRJCx',
