+++ packages/styleguide/stories/Molecules/Popover/index.stories.mdx
@@ -18,7 +18,7 @@ import {
   parameters={{
     subtitle: `Popover is a surface which takes a target element and shows an overlay
 positioned around that element.`,
-    source: 'gamut-labs',
+    source: 'gamut',
     status: 'updating',
   }}
 />
